hw/input: cleanup and cosmetics
