+++ hw/input/game_controller_test.go
@@ -12,14 +12,14 @@ func TestInputCodeMarshalRoundTrip(t *testing.T) {
 		text string
 		code *Code // nil for unmarsal errors
 	}{
-		{"", &Code{Type: UnsetController}},
-		{"key W", &Code{Type: Keyboard, Scancode: sdl.SCANCODE_W}},
-		{"key Up", &Code{Type: Keyboard, Scancode: sdl.SCANCODE_UP}},
-		{"key Return", &Code{Type: Keyboard, Scancode: sdl.SCANCODE_RETURN}},
-		{"joybtn a 030000004c050000cc0900", &Code{Type: ControllerButton, CtrlButton: sdl.CONTROLLER_BUTTON_A, CtrlGUID: "030000004c050000cc0900"}},
-		{"joybtn x 030000004c050000cc0900", &Code{Type: ControllerButton, CtrlButton: sdl.CONTROLLER_BUTTON_X, CtrlGUID: "030000004c050000cc0900"}},
-		{"joyaxis righttrigger+ 030000004c050000cc1212", &Code{Type: ControllerAxis, CtrlAxis: sdl.CONTROLLER_AXIS_TRIGGERRIGHT, CtrlAxisDir: 1, CtrlGUID: "030000004c050000cc1212"}},
-		{"joyaxis lefttrigger- 123400004c050000cc1212", &Code{Type: ControllerAxis, CtrlAxis: sdl.CONTROLLER_AXIS_TRIGGERLEFT, CtrlAxisDir: -1, CtrlGUID: "123400004c050000cc1212"}},
+		{"", &Code{Type: ControlNotSet}},
+		{"key W", &Code{Type: KeyboardCtrl, Scancode: sdl.SCANCODE_W}},
+		{"key Up", &Code{Type: KeyboardCtrl, Scancode: sdl.SCANCODE_UP}},
+		{"key Return", &Code{Type: KeyboardCtrl, Scancode: sdl.SCANCODE_RETURN}},
+		{"joybtn a 030000004c050000cc0900", &Code{Type: ButtonCtrl, CtrlButton: sdl.CONTROLLER_BUTTON_A, CtrlGUID: "030000004c050000cc0900"}},
+		{"joybtn x 030000004c050000cc0900", &Code{Type: ButtonCtrl, CtrlButton: sdl.CONTROLLER_BUTTON_X, CtrlGUID: "030000004c050000cc0900"}},
+		{"joyaxis righttrigger+ 030000004c050000cc1212", &Code{Type: AxisCtrl, CtrlAxis: sdl.CONTROLLER_AXIS_TRIGGERRIGHT, CtrlAxisDir: 1, CtrlGUID: "030000004c050000cc1212"}},
+		{"joyaxis lefttrigger- 123400004c050000cc1212", &Code{Type: AxisCtrl, CtrlAxis: sdl.CONTROLLER_AXIS_TRIGGERLEFT, CtrlAxisDir: -1, CtrlGUID: "123400004c050000cc1212"}},
 
 		// unmarsal errors
 		{"key   ", nil},
